Refactor the dialog to confirm protocol
Close #144
